Update 321db shim with same changes as 31xdb shim, including moving meta class to separate file and fixing issue in pre-merge build
