+++ sql-plugin/src/main/321db/scala/com/nvidia/spark/rapids/shims/SparkShims.scala
@@ -27,18 +27,17 @@ import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions._
-import org.apache.spark.sql.catalyst.plans.physical.{BroadcastMode, IdentityBroadcastMode}
+import org.apache.spark.sql.catalyst.plans.physical.BroadcastMode
 import org.apache.spark.sql.execution._
 import org.apache.spark.sql.execution.adaptive.{BroadcastQueryStageExec, ShuffleQueryStageExec}
 import org.apache.spark.sql.execution.datasources._
 import org.apache.spark.sql.execution.datasources.json.JsonFileFormat
-import org.apache.spark.sql.execution.exchange.{BroadcastExchangeExec, ReusedExchangeExec}
-import org.apache.spark.sql.execution.joins.HashedRelationBroadcastMode
+import org.apache.spark.sql.execution.exchange.ReusedExchangeExec
 import org.apache.spark.sql.execution.python.WindowInPandasExec
-import org.apache.spark.sql.execution.reuse.ReuseExchangeAndSubquery
 import org.apache.spark.sql.execution.window.WindowExecBase
 import org.apache.spark.sql.rapids.GpuFileSourceScanExec
-import org.apache.spark.sql.rapids.execution.{GpuBroadcastExchangeExec, GpuBroadcastMeta, GpuSubqueryBroadcastExec}
+import org.apache.spark.sql.rapids.execution.{GpuBroadcastExchangeExec, GpuSubqueryBroadcastExec}
+import org.apache.spark.sql.rapids.execution.shims.{GpuSubqueryBroadcastMeta, ReuseGpuBroadcastExchangeAndSubquery}
 import org.apache.spark.sql.rapids.shims.GpuFileScanRDD
 import org.apache.spark.sql.types._
 
@@ -100,7 +99,7 @@ object SparkShimImpl extends Spark321PlusShims with Spark320until340Shims {
 
   override def applyPostShimPlanRules(plan: SparkPlan): SparkPlan = {
     val rules = Seq(
-      ReuseExchangeAndSubquery
+      ReuseGpuBroadcastExchangeAndSubquery
     )
     rules.foldLeft(plan) { case (sp, rule) =>
       rule.apply(sp)
@@ -112,97 +111,7 @@ object SparkShimImpl extends Spark321PlusShims with Spark320until340Shims {
       GpuOverrides.exec[SubqueryBroadcastExec](
         "Plan to collect and transform the broadcast key values",
         ExecChecks(TypeSig.all, TypeSig.all),
-        (s, conf, p, r) => new SparkPlanMeta[SubqueryBroadcastExec](s, conf, p, r) {
-          private var broadcastBuilder: () => SparkPlan = _
-
-          override val childExprs: Seq[BaseExprMeta[_]] = Nil
-
-          override val childPlans: Seq[SparkPlanMeta[SparkPlan]] = Nil
-
-          override def tagPlanForGpu(): Unit = s.child match {
-            // DPP: For AQE off, in this case, we handle DPP by converting the underlying 
-            // BroadcastExchangeExec to GpuBroadcastExchangeExec
-            // This is slightly different from the Apache Spark case, because Spark 
-            // sends the underlying plan into the plugin in advance via the PlanSubqueries rule
-            // here, we have the full non-GPU subquery plan, so we convert the whole
-            // thing here.
-            case ex @ BroadcastExchangeExec(_, child) =>
-              val exMeta = new GpuBroadcastMeta(ex.copy(child = child), conf, p, r)
-              exMeta.tagForGpu()
-              if (exMeta.canThisBeReplaced) {
-                broadcastBuilder = () => exMeta.convertToGpu()
-              } else {
-                willNotWorkOnGpu("underlying BroadcastExchange can not run in the GPU.")
-              }
-            // DPP: For AQE on, we have an almost completely different scenario then before, 
-            // Databricks uses a BroadcastQueryStageExec and either:
-            //  1) provide an underlying BroadcastExchangeExec that we will have to convert
-            //     somehow
-            //  2) might already do the reuse work for us. The ReusedExchange is now a 
-            //     part of the SubqueryBroadcast, so we send it back here as underlying the 
-            //     GpuSubqueryBroadcastExchangeExec
-            case bqse: BroadcastQueryStageExec =>
-              bqse.plan match {
-                case ex: BroadcastExchangeExec =>
-                  val exMeta = new GpuBroadcastMeta(ex, conf, p, r)
-                  exMeta.tagForGpu()
-                  if (exMeta.canThisBeReplaced) {
-                    broadcastBuilder = () => exMeta.convertToGpu()
-                    // broadcastBuilder = () =>
-                    //   SparkShimImpl.columnarAdaptivePlan(
-                    //     a, TargetSize(conf.gpuTargetBatchSizeBytes))
-                  } else {
-                    willNotWorkOnGpu("underlying BroadcastExchange can not run in the GPU.")
-                  }
-                case reuse: ReusedExchangeExec =>
-                  reuse.child match {
-                    case gpuExchange: GpuBroadcastExchangeExec =>
-                      // A BroadcastExchange has already been replaced, so it can run on the GPU
-                      broadcastBuilder = () => reuse
-                    case _ =>
-                      willNotWorkOnGpu("underlying BroadcastExchange can not run in the GPU.")
-                  }
-              }
-            case _ =>
-              willNotWorkOnGpu("the subquery to broadcast can not entirely run in the GPU.")
-          }
-          /**
-          * Simply returns the original plan. Because its only child, BroadcastExchange, doesn't
-          * need to change if SubqueryBroadcastExec falls back to the CPU.
-          */
-          override def convertToCpu(): SparkPlan = s
-
-          override def convertToGpu(): GpuExec = {
-            GpuSubqueryBroadcastExec(s.name, s.index, s.buildKeys, broadcastBuilder())(
-              getBroadcastModeKeyExprs)
-          }
-
-          /** Extract the broadcast mode key expressions if there are any. */
-          private def getBroadcastModeKeyExprs: Option[Seq[Expression]] = {
-            val broadcastMode = s.child match {
-              case b: BroadcastExchangeExec =>
-                b.mode
-              case bqse: BroadcastQueryStageExec =>
-                bqse.plan match {
-                  case b: BroadcastExchangeExec =>
-                    b.mode
-                  case reuse: ReusedExchangeExec =>
-                    reuse.child match {
-                      case g: GpuBroadcastExchangeExec =>
-                        g.mode
-                    }
-                  case _ =>
-                    throw new AssertionError("should not reach here")
-                }
-            }
-
-            broadcastMode match {
-              case HashedRelationBroadcastMode(keys, _) => Some(keys)
-              case IdentityBroadcastMode => None
-              case m => throw new UnsupportedOperationException(s"Unknown broadcast mode $m")
-            }
-          }
-        }),
+        (s, conf, p, r) => new GpuSubqueryBroadcastMeta(s, conf, p, r)),
       GpuOverrides.exec[FileSourceScanExec](
         "Reading data from files, often from Hive tables",
         ExecChecks((TypeSig.commonCudfTypes + TypeSig.NULL + TypeSig.STRUCT + TypeSig.MAP +
